requirements.txt
torchvision
numpy
tqdm
tensorboard
matplotlib

pandas
datasets
librosa
wandb